refactor zip import logic into service
